refactor: extract shared test infrastructure from practice tests
Move FakeCursor, FakeCollection, FakeDatabase, matches_query,
make_user, and make_problem to backend/tests/api/conftest.py.
Both test_practice.py and test_practice_submit.py now import
from the shared module instead of duplicating definitions.

Fixes #30
